fix: resolve single-instance test race; clear clippy warnings
- process: pass data_dir explicitly to pid_path/ensure_single_instance/
  write_pid_file/stop instead of reading global REX_DATA_DIR; removes a
  parallel-test race on REX_DATA_DIR and a hidden global-env dependence
- process test: pass explicit dir, no global env mutation
- clippy: drop redundant matches! in cli test; remove usize->usize cast in
  db.rs; widen agent_ws payload visibility to pub(crate); fix rex-sip
  build.rs doc indentation + useless format!

+++ crates/rex-hub/src/db.rs
@@ -678,7 +678,7 @@ impl Database {
                     rusqlite::params![old, old],
                 )
                 .map_err(|e| RExError::Message(e.to_string()))?;
-            count += n as usize;
+            count += n;
         }
         Ok(count)
     }
